Changes on RESTORE_ON_START behaviour
If the startup parameter start is used and RESTORE_ON_START is
enabled and if an error occurs, while loading the iptables rules,
IP-Array will no longer try to start regulary.
Also there are more informative messages about the loading process
of the prepared scripts.
Added a warning message, that dry-run mode is ignore with this
parameter combination.